+++ src/pages-xpto/home/components/OfferCard.tsx
