+++ src/pages-xpto/home/components/SearchBar.tsx
