+++ src/pages-xpto/home/home.tsx
