+++ src/pages-xpto/offer/index.ts
