+++ src/pages-xpto/tmp_xpto/steps/index.ts
