--- src/pages/home/components/CreateAdPopUp.tsx
@@ -1,44 +1,0 @@
-import { Link } from 'react-router-dom'
-
-export const CreateAdPopUp = () => {
-  return (
-    <div className='relative bg-white w-full max-w-md p-8'>
-      <div className='flex justify-end mb-8'>
-        <img
-          src='kumbulink-logo.svg'
-          alt='Kumbulink Logo'
-          className='w-40 h-40'
-        />
-      </div>
-
-      <div className='space-y-4'>
-        <h2 className='text-[30px] leading-tight'>
-          <span className='text-primary-orange font-normal'>JUNTA-TE À</span>
-          <br />
-          <span className='font-bold'>KUMBULINK</span>
-        </h2>
-
-        <p className='text-[16px] leading-7 text-gray-700 max-w-[280px]'>
-          Regista-te agora e encontra milhares de pessoas para trocar e enviar
-          dinheiro de forma segura, simples e a partir de qualquer lugar.
-        </p>
-
-        <div className='pt-4'>
-          <Link
-            to='/registrar'
-            className='block w-full bg-primary-green text-white py-3 rounded text-center font-medium text-lg hover:bg-primary-green/90'
-          >
-            Registar
-          </Link>
-
-          <Link
-            to='/login'
-            className='block text-primary-orange text-center mt-4 text-sm hover:underline'
-          >
-            Já tenho uma conta
-          </Link>
-        </div>
-      </div>
-    </div>
-  )
-}
